encapsulate canvas logic
