+++ packages/widgets/index.mjs
@@ -1,2 +1,3 @@
 export * from './Claviature.jsx';
 export * from './Pianoroll.jsx';
+export * from './Canvas.jsx';
